requirements.txt
cryptography>=42.0.5
PyJWT>=2.8.0
scikit-learn>=1.4.2
numpy>=2.0.0
